+++ core/gen_data/gen_data.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 
 	. "github.com/candid82/joker/core"
+	. "github.com/candid82/joker/core/gen_common"
 	_ "github.com/candid82/joker/std/html"
 	_ "github.com/candid82/joker/std/string"
 )
